commit message

+++ src/main/java/com/bway/springdemo/repository/EmployeeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.bway.springdemo.model.Employee;
 import com.bway.springdemo.model.User;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 
 }
